refactor: move all assets to public folder & reconfigure webpack.config
