+++ .editorconfig
@@ -4,6 +4,7 @@ root = true
 charset = utf-8
 indent_style = space
 indent_size = 2
+tab_width = 2
 end_of_line = lf
 trim_trailing_whitespace = true
 insert_final_newline = true
