+++ package.json
@@ -3,16 +3,17 @@
   "version": "0.0.0",
   "private": true,
   "main": "index.js",
-  "author": "vladKB1, Tsvetinskaya-L, AntonKos",
+  "author": "vladKB1",
   "license": "MIT",
   "scripts": {
     "build": "webpack",
     "start": "webpack serve --env development",
-    "lint": "eslint src"
+    "lint": "eslint src",
+    "dev": "webpack --env development"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^5.10.2",
-    "@typescript-eslint/parser": "^5.10.2",
+    "@typescript-eslint/parser": "^5.24.0",
     "bootstrap": "^5.1.3",
     "bootstrap-icons": "^1.8.1",
     "clean-webpack-plugin": "^4.0.0",
@@ -21,7 +22,9 @@
     "eslint": "^8.8.0",
     "eslint-config-airbnb-typescript": "^16.1.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.25.4",
+    "eslint-import-resolver-typescript": "^2.7.1",
+    "eslint-import-resolver-webpack": "^0.13.2",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-webpack-plugin": "^3.1.1",
     "html-loader": "^3.1.0",
@@ -33,7 +36,7 @@
     "sass-loader": "^12.4.0",
     "sass-resources-loader": "^2.2.4",
     "ts-loader": "^9.2.6",
-    "typescript": "^4.5.5",
+    "typescript": "^4.6.4",
     "webpack": "^5.68.0",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.7.4"
