+++ src/index.html
@@ -4,7 +4,8 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title></title>
+    <link rel="shortcut icon" href="../public/favicon.ico" type="image/x-icon" />
+    <title>Rs lang</title>
   </head>
   <body></body>
 </html>
