+++ src/index.ts
@@ -1,13 +1,6 @@
 import App from './App';
 import './styles/main.scss';
 
-import './assets/images/sprint/icons/soundOn.svg';
-import './assets/images/sprint/icons/soundOff.svg';
-import './assets/images/sprint/icons/sound-on_121151.svg';
-import './assets/images/sprint/icons/home.svg';
-import './assets/images/sprint/icons/premium-icon-sound-3293595.png';
-import './assets/images/png/icon-textbook.png';
-
 const app = new App();
 document.addEventListener('DOMContentLoaded', () => {
   app.init();
