+++ src/styles/layout/_sprint.scss
@@ -20,7 +20,7 @@
   display: none;
   background: none;
   width: 50px;
-  background-image: url('../assets/images/sprint/icons/soundOn.svg');
+  background-image: url('../../public/assets/images/sprint/icons/soundOn.svg');
   background-repeat: no-repeat;
 }
 
@@ -29,7 +29,7 @@
   border: none;
   background: none;
   width: 50px;
-  background-image: url('../assets/images/sprint/icons/home.svg');
+  background-image: url('../../public/assets/images/sprint/icons/home.svg');
   background-repeat: no-repeat;
 }
 
@@ -198,7 +198,7 @@
   width: 40px;
   height: 40px;
   background-repeat: no-repeat;
-  background-image: url('../assets/images/sprint/icons/sound-on_121151.svg');
+  background-image: url('../../public/assets/images/sprint/icons/sound-on_121151.svg');
 }
 
 .icon-sound {
